Add support for newinstall.sh's Conda environment.
This is necessary to be able to bootstrap a stack using the current
newinstall.sh.

This forces the name of the Conda environment to something simple (rather than
trying to derive it from a SHA1, as newinstall.sh does by default).

It sets a bunch of environment variables to look like we're in the Conda
environment when installing packages. This isn't as robust as properly
activating the environment, but it's easy to implement...!